project_langchain/requirements.txt
# Core dependencies for LangChain RAG Pipeline

# LangChain packages
langchain>=0.1.0
langchain-community>=0.0.20
langchain-chroma>=0.1.0
langchain-huggingface>=0.0.1

# NLP and embeddings
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0
accelerate>=0.20.0

# Vector database
chromadb>=0.4.0

# UI
gradio>=4.0.0

# Configuration
pyyaml>=6.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0

